refactor: update request and response headers to use new headers package
- Replaced instances of kafkaapi.ResponseHeader and kafkaapi.RequestHeader with headers.ResponseHeader and headers.RequestHeader across multiple files, improving code organization and clarity.
- Introduced a new headers package to encapsulate request and response header definitions, enhancing modularity and maintainability of the codebase.
