+++ internal/stage_c1.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/codecrafters-io/kafka-tester/internal/kafka_executable"
 	kafkaapi "github.com/codecrafters-io/kafka-tester/protocol/api"
+	headers "github.com/codecrafters-io/kafka-tester/protocol/api/headers"
 	"github.com/codecrafters-io/kafka-tester/protocol/kafka_client"
 	"github.com/codecrafters-io/kafka-tester/protocol/serializer"
 	"github.com/codecrafters-io/tester-utils/random"
@@ -37,7 +38,7 @@ func testSequentialRequests(stageHarness *test_case_harness.TestCaseHarness) err
 	for i := range requestCount {
 		correlationId := getRandomCorrelationId()
 		request := kafkaapi.ApiVersionsRequest{
-			Header: kafkaapi.RequestHeader{
+			Header: headers.RequestHeader{
 				ApiKey:        18,
 				ApiVersion:    4,
 				CorrelationId: correlationId,
