+++ internal/stage_c2.go
@@ -7,6 +7,7 @@ import (
 	"github.com/codecrafters-io/kafka-tester/internal/kafka_executable"
 	"github.com/codecrafters-io/kafka-tester/protocol"
 	kafkaapi "github.com/codecrafters-io/kafka-tester/protocol/api"
+	headers "github.com/codecrafters-io/kafka-tester/protocol/api/headers"
 	"github.com/codecrafters-io/kafka-tester/protocol/kafka_client"
 	"github.com/codecrafters-io/kafka-tester/protocol/serializer"
 	"github.com/codecrafters-io/tester-utils/logger"
@@ -48,7 +49,7 @@ func testConcurrentRequests(stageHarness *test_case_harness.TestCaseHarness) err
 	for i, client := range clients {
 		correlationIds[i] = int32(random.RandomInt(-math.MaxInt32, math.MaxInt32))
 		request := kafkaapi.ApiVersionsRequest{
-			Header: kafkaapi.RequestHeader{
+			Header: headers.RequestHeader{
 				ApiKey:        18,
 				ApiVersion:    4,
 				CorrelationId: correlationIds[i],
